Fix Bash glob when Rust hash ends in d
When the Rust hash ends in a "d" the previous blob would fail to match.
This is due to #3 which fixed an issue where Rust generated ".d" files
that should correctly be filtered.

This new version loops over both the correct files and ".d" files.
Within the loop it checks for if the file is executable as its
filtering of the ".d" files.

I believe in general this is a better design as the more advanced Bash
glob is not very portable.